Factor out ImageDecoration and paintImage
To make it cleaner when we RTLify these.
